Fixed !queue Moved some startup items to startup.cs did some work on the cache folder
