+++ discordMusicBot/discordMusicBot.csproj
@@ -102,6 +102,7 @@
     <Compile Include="src\network.cs" />
     <Compile Include="src\player.cs" />
     <Compile Include="src\playlist.cs" />
+    <Compile Include="src\startup.cs" />
     <Compile Include="src\system.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -116,6 +117,7 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
