+++ discordMusicBot/src/player.cs
@@ -84,7 +84,6 @@ namespace discordMusicBot.src
 
                             }
                             
-
                         }
                     }
                 }
